Cleanup spec to improve package structure
