+++ src/main/kotlin/dev/themeinerlp/skinserver/handler/head/HeadHandlerImpl.kt
@@ -5,7 +5,7 @@ import dev.themeinerlp.skinserver.service.GameProfileService
 import dev.themeinerlp.skinserver.service.RenderService
 import dev.themeinerlp.skinserver.service.SkinService
 import dev.themeinerlp.skinserver.spec.dao.skin.Skin
-import dev.themeinerlp.skinserver.spec.handler.head.HeadDatabaseHandler
+import dev.themeinerlp.skinserver.spec.database.HeadDatabaseHandler
 import dev.themeinerlp.skinserver.spec.repository.skin.SkinRepository
 import dev.themeinerlp.skinserver.utils.HeadView
 import java.util.Base64
