+++ src/main/kotlin/dev/themeinerlp/skinserver/spec/database/RenderDatabaseHandler.kt
@@ -1,0 +1,9 @@
+package dev.themeinerlp.skinserver.spec.database
+
+import dev.themeinerlp.skinserver.utils.HeadView
+import org.springframework.http.ResponseEntity
+import org.springframework.web.multipart.MultipartFile
+
+interface RenderDatabaseHandler {
+    fun renderHead(skin: MultipartFile?, size: Int?, rotation: HeadView?, layer: Boolean?): ResponseEntity<Any>
+}
